+++ tests/test_pixel_level_inputs.py
@@ -244,7 +244,6 @@ def test_generate_marker_correlation_information():
     assert children
     assert len(children) > 5
 
-
 def test_blank_reset_histogram():
     blank_hist = reset_pixel_histogram(True)
     assert blank_hist['layout']['margin'] == {'b': 15, 'l': 5, 'pad': 0, 'r': 5, 't': 20}